requirements.txt
asgiref==3.2.10
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
django-import-export==2.4.0
et-xmlfile==1.0.1
folium==0.12.1
gmplot==1.4.1
googlemaps==4.4.2
gpxpy==1.4.2
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.2.2
Pillow==8.1.0
psycopg2==2.8.6
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
Shapely==1.7.1
six==1.15.0
sodapy==2.1.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.3
whitenoise==5.2.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
